Updated Database
Implemented MongoDb Stitch to login anonymously if User has already registered. If not they need to create an account, then it will register that User with that specific device. Does not login to an account on another device even if that account was made on a previous one. It doesn't seem that tatical, but for this implemtation, it's fine. Also created an Observer that notifies the fragment when the Database has finished its tasks and then updates the rest of the information.