Add user to data base (#2)
* Created MongoDB

Added functioality to connect to DB online and push docs

* Updated Database

Implemented MongoDb Stitch to login anonymously if User has already registered. If not they need to create an account, then it will register that User with that specific device. Does not login to an account on another device even if that account was made on a previous one. It doesn't seem that tatical, but for this implemtation, it's fine. Also created an Observer that notifies the fragment when the Database has finished its tasks and then updates the rest of the information.

* Removed Function in UserLab

Removed a function in the UserLab that doesn't have a user at the current moment. Will implement again soon.

* Updated Login Credentials

Made it so no spaces are allowed in the password or username. Also, made the password length so that it must be greater than 5

* Reformated Code

Made the code stay within the lines.