Add mobile vision api (#3)
* Mobile Vision Implemented

Implemented the Google Mobile Vision Api. Opens the camera, then asks for permission. Take a picture to scan the barcode. Doesn't add anything to a database yet, but the barcode is retrieved.

* Points For Barcode Added

Added the ability to get points for barcodes scanned. If there is already that barcode in the database it is not added and points are not added.