Leader board (#4)
* Added the LeaderBoard

Added the Recycler view for the LeaderBoard. Also fixed a bug in the camera that would result in a possible crash, when the user is first prompted for the camera permissions. LeaderBoard is randomized, with a preset list of names (Just common ones I found on the internet). The users name is factored in aswell.

* Added Special Users to LeaderBoards

Added some special entries to the LeaderBoards. Some top Recyclers over here.

* Update BarcodeCameraActivity.java

Removed unused import